+++ R/plot_macrosynteny.R
@@ -24,8 +24,8 @@
 #' my_macrosynteny <- compute_macrosynteny(my_orthologs)
 #' 
 #' plot_macrosynteny(my_macrosynteny,
-#'                   sp1_label = "B.floridae",
-#'                   sp2_label = "P.echinospica")
+#'                   sp1_label = "B. floridae",
+#'                   sp2_label = "P. yessoensis")
 #' 
 #' @export