+++ vignettes/macrosyntR.R
@@ -10,8 +10,18 @@ library(macrosyntR)
 ## -----------------------------------------------------------------------------
 
 my_orthologs_table <- load_orthologs(orthologs_table = system.file("extdata","Bflo_vs_Pech.tab",package="macrosyntR"),
-                                     bedfiles = c(system.file("extdata","Bflo.protein_products.bed",package="macrosyntR"),
-                                     system.file("extdata","Pech.protein_products.bed",package="macrosyntR")))
+                                     bedfiles = c(system.file("extdata","Bflo.bed",package="macrosyntR"),
+                                     system.file("extdata","Pech.bed",package="macrosyntR")))
+
+head(my_orthologs_table)
+
+
+## -----------------------------------------------------------------------------
+
+my_orthologs_with_3_sp <- load_orthologs(orthologs_table = system.file("extdata","Single_copy_orthologs.tsv",package="macrosyntR"),
+                                     bedfiles = c(system.file("extdata","Bflo.bed",package="macrosyntR"),
+                                                  system.file("extdata","Pech.bed",package="macrosyntR"),
+                                                  system.file("extdata","Pyes.bed",package="macrosyntR")))
 
 head(my_orthologs_table)